+++ src/ccmain/pageiterator.cpp
@@ -495,9 +495,9 @@ bool PageIterator::Baseline(PageIteratorLevel level,
            ? word->bounding_box()
            : row->bounding_box();
   int left = box.left();
-  ICOORD startpt(left, static_cast<int16_t>(row->base_line(left) + 0.5));
+  ICOORD startpt(left, static_cast<int16_t>(row->base_line(static_cast<float>(left)) + 0.5f));
   int right = box.right();
-  ICOORD endpt(right, static_cast<int16_t>(row->base_line(right) + 0.5));
+  ICOORD endpt(right, static_cast<int16_t>(row->base_line(static_cast<float>(right)) + 0.5));
   // Rotate to image coordinates and convert to global image coords.
   startpt.rotate(it_->block()->block->re_rotation());
   endpt.rotate(it_->block()->block->re_rotation());
